Split projectController
