+++ frontend/src/core/models/const-resources/toastr-config.ts
